rework of weighted graphs API, fixes #6 and Graphs#42 (#14)
I removed all methods using internal fields from the abstract functions.
I fixed the issue #6 occurring when removing an edge. It now delete the
coefficient (set the coefficient to a structural zero), instead of just
zeroing it. It is done more efficiently than by calling dropzeros! and
is as efficient as adding a new edge (outside for modifying a weight).
Now, the non edges are always structural zeros, so technically, we could
allow setting zero weight edges. I did not permit that as it would
probably be breaking, and because we can't add zeros from the weights
matrix constructor, as zeros are mapped to structural zeros.

Review needed please.

---------

Co-authored-by: Guillaume Dalle <22795598+gdalle@users.noreply.github.com>
Co-authored-by: Simon Schölly <sischoel@gmail.com>